Clear logins when logins out from FxA
